chore: resolve pr feedback
